requirements.txt
alpaca-trade-api==2.3.0
streamlit==1.12.0
backtrader==1.9.76.123
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.5
matplotlib==3.5.1
plotly==5.5.0
joblib==1.1.0
requests==2.26.0
python-dotenv==0.19.2
ruff==0.0.100  # Add Ruff for linting
